% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_seatbelt}
\alias{accident_seatbelt}
\title{# 汽车事故与安全带 (Auto Accidents and Seat Belts)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2 x 5.}
\usage{
data(accident_seatbelt)
}
\description{
# 汽车事故与安全带 (Auto Accidents and Seat Belts)
}
\keyword{datasets}
