% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deathpenalty}
\alias{deathpenalty}
\title{死刑判决1 (Death Penalty I)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2.}
\usage{
data(deathpenalty)
}
\description{
死刑判决1 (Death Penalty I)
}
\keyword{datasets}
