% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{find_data_by_title}
\alias{find_data_by_title}
\alias{find_data_by_var}
\title{Find a dataset by dataset description title or dataset variable name}
\usage{
find_data_by_title(pattern, package = "xmucdaDB", ignore.case = TRUE,
  perl = FALSE, fixed = FALSE, simplify = TRUE)

find_data_by_var(pattern, package = "xmucdaDB", ignore.case = TRUE,
  perl = FALSE, fixed = FALSE, simplify = TRUE)
}
\arguments{
\item{pattern, ignore.case, perl, fixed}{We use \code{grep} to match
  the pattern with dataset description or with variable name.
  All those parameters that would be passed to \code{grep},
  see \code{\link{grep}} for more information.

  Note that the default value of \code{ignore.case} in \code{grep} is FALSE, but
  in this two functions, the default value is TRUE.}

\item{package}{a character vector giving the package(s) to look in for data sets,
  or NULL (for looking for dataset in all data sets).

  By default, this two functions would find dataset in \code{xmucdaDB} package.}

\item{simplify}{a logical. If TRUE (default), function would return
the dataset name only. If FALSE, it would return a \code{data.frame}
which includes the package, dataset name, dataset description and
variable name that match the pattern (only for \code{find_data_by_var}).}
}
\description{
Find the required data sets quickly when the dataset name is unknown
but some description about the dataset or some variable name is known.
}
\examples{
# 描述中包含"癌"的数据集
find_data_by_title("癌")
find_data_by_title("癌", simplify = FALSE)

# 描述中包含"癌"且有变量名包含"treatment"的数据集
intersect(
  find_data_by_title("癌"),
  find_data_by_var("treatment", ignore.case = TRUE)
)

}
