% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{job_satisfaction_survey}
\alias{job_satisfaction_survey}
\title{工作满意度与种族、性别、年龄、工作区域的关系的调查 (Survey on  How Job Satisfaction Depends on Race, Gender, Age, and Regional Location)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 3 x 7 x 2 x 2.}
\usage{
data(job_satisfaction_survey)
}
\description{
工作满意度与种族、性别、年龄、工作区域的关系的调查 (Survey on  How Job Satisfaction Depends on Race, Gender, Age, and Regional Location)
}
\keyword{datasets}
