% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MBdrink}
\alias{MBdrink}
\title{饮酒与Meyers-Briggs人格测验 (Drinking and the Meyers-Briggs Personality Test)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2 x 2 x 2.}
\usage{
data(MBdrink)
}
\description{
饮酒与Meyers-Briggs人格测验 (Drinking and the Meyers-Briggs Personality Test)
}
\keyword{datasets}
