% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premarital_sex}
\alias{premarital_sex}
\title{婚前性行为 (Premarital Sex)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 3 x 2 x 2.}
\usage{
data(premarital_sex)
}
\description{
婚前性行为 (Premarital Sex)
}
\keyword{datasets}
