% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatment}
\alias{treatment}
\title{治疗中心、治疗方案与响应 (Clinic, Treatment and Response)}
\format{An object of class \code{xtabs} (inherits from \code{table}) of dimension 2 x 2 x 2.}
\usage{
data(treatment)
}
\description{
治疗中心、治疗方案与响应 (Clinic, Treatment and Response)
}
\keyword{datasets}
