% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmucdaDB-package.R
\docType{package}
\name{xmucdaDB}
\alias{xmucdaDB}
\alias{xmucdaDB-package}
\title{Data of "An Introduction to Categorical Data Analysis,
  2nd edition by Alan Agresti"}
\description{
Data of "An Introduction to Categorical Data Analysis,
  2nd edition by Alan Agresti".

  Some of data files in this packages are from package "icda" by Brett Presnell
  (http://users.stat.ufl.edu/~presnell/Courses/sta4504-2012sp/R-info.shtml).
}
\details{
xmucdaDB.


This package include data sets that may be used in the textbook.
Using \code{data(package = "xmucdaDB")} to see all the data sets
that are avaliable in this package.

It also include two functions (\code{find_data_by_title} and
\code{find_data_by_var}) to make it convenient to find required dataset
from so many data sets in this package. See \code{\link{find_data_by_title}}.
}
